// ui/src/@types/TableWidget.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Unique identifier for the table widget
 */
export type Id = string;
/**
 * Unique identifier for the row
 */
export type Id1 = string;
/**
 * Display value of the row
 */
export type DisplayValue = string | null;
/**
 * Corpus id of a paper
 */
export type PaperCorpusId = number | null;
/**
 * List of rows in the table
 */
export type Rows = TableRow[];
/**
 * Unique identifier for the column
 */
export type Id2 = string;
/**
 * Name of the column
 */
export type Name = string;
/**
 * Description of the column
 */
export type Description = string;
/**
 * Flag indicating whether the column to be populated is a metadata column or not
 */
export type IsMetadata = boolean | null;
/**
 * List of tools used to generate the column values
 */
export type Tools = string[];
/**
 * List of columns in the table
 */
export type Columns = TableColumn[];
/**
 * Unique identifier for the cell. In the format ${row.id}-${column.id}
 */
export type Id3 = string;
/**
 * Value of the cell
 */
export type Value = string | number | null;
/**
 * Display value of the cell
 */
export type DisplayValue1 = string;
/**
 * Whether the cell is still loading
 */
export type IsLoading = boolean;
/**
 * Error message if any
 */
export type Error = string | null;
/**
 * Metadata of the cell
 */
export type Metadata = {
  [k: string]: unknown;
} | null;
/**
 * overridden cell value by a user, nora, or other 3rd party
 */
export type EditedValue = string | null;
/**
 * Who last edited the table cell
 */
export type EditedByUuid = string | null;
/**
 * Title for table generated
 */
export type Title = string | null;

export interface TableWidget {
  id: Id;
  rows?: Rows;
  columns?: Columns;
  cells?: Cells;
  title?: Title;
  [k: string]: unknown;
}
export interface TableRow {
  id: Id1;
  display_value?: DisplayValue;
  paper_corpus_id: PaperCorpusId;
  [k: string]: unknown;
}
export interface TableColumn {
  id: Id2;
  name: Name;
  description: Description;
  is_metadata?: IsMetadata;
  tools: Tools;
  [k: string]: unknown;
}
/**
 * Dictionary of cells in the table. Key is in the format ${row.id}-${column.id}
 */
export interface Cells {
  [k: string]: TableCell;
}
export interface TableCell {
  id: Id3;
  value?: Value;
  display_value: DisplayValue1;
  is_loading?: IsLoading;
  error?: Error;
  metadata?: Metadata;
  edited_value?: EditedValue;
  edited_by_uuid?: EditedByUuid;
  [k: string]: unknown;
}